requirements.txt
boto3>=1.34
pyyaml>=6.0
psycopg2-binary>=2.9
pymysql>=1.1
